Use private copy of memmem() to find the UUID
The current implementation of the routine which searches for the
LTN UUID SEI message is quite expensive, and doesn't scale well to
streams that are 200Mbps+.

Switch to using memmem() to find the byte array.  While different
libc versions include a version of this function (e.g. glibc has
it if GNU_SOURCE is defined), using our own version ensures the
function is available (as it's not part of the POSIX standard), and
that the performance of the implementation is consistent regardless
of platform.