Fix RTMP clockWrapValue
The code most likely wanted a power of two, not 2 xor 32.

This isn't entirely correct yet: RTMP specifies that two timestamps with
a uint32 delta of more than 2^31-1 ms are regressing, not wrapping, so
assuming we wrapped when we see a regression is only correct when
timestamps are monotonic (which they usually are).