smoother-pcr: Record the first PCR walltime sooner
The code waited for a second PCR before writing out the entire interval.
However, receive timestamping happened after the wait, leading to the
scheduled output timestamps being off by a PCR interval.

We can correct at least the scheduling by timestamping the first PCR as
soon as we see it.